+++ config/config.js
@@ -1,0 +1,19 @@
+export const continents = ["africa", "americas", "asia", "europe", "oceania"];
+
+export const calculationMethods = [
+  { id: 1, name: "University of Islamic Sciences, Karachi" },
+  { id: 2, name: "Islamic Society of North America (ISNA)" },
+  { id: 3, name: "Muslim World League (MWL)" },
+  { id: 4, name: "Umm al-Qura, Makkah" },
+  { id: 5, name: "Egyptian General Authority of Survey" },
+  { id: 7, name: "Institute of Geophysics, University of Tehran" },
+  { id: 8, name: "Gulf Region" },
+  { id: 9, name: "Kuwait" },
+  { id: 10, name: "Qatar" },
+  { id: 11, name: "Majlis Ugama Islam Singapura, Singapore" },
+  { id: 12, name: "Union Organization islamic de France" },
+  { id: 13, name: "Diyanet İşleri Başkanlığı, Turkey" },
+  { id: 14, name: "Spiritual Administration of Muslims of Russia" },
+];
+
+export const API_BASE_URL = "https://api.aladhan.com/v1";
