+++ index.html
@@ -9,6 +9,7 @@
       href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.4.0/css/all.min.css"
     />
     <link rel="stylesheet" href="style.css" />
+    <script type="module" src="app/app.js"></script>
   </head>
   <body>
     <main>
@@ -87,6 +88,6 @@
         </div>
       </section>
     </main>
-    <script type="module" src="script.js"></script>
+    
   </body>
 </html>
